+++ lib/src/enums/player_state.dart
@@ -6,12 +6,29 @@
 ///
 /// Find more about it [here](https://developers.google.com/youtube/iframe_api_reference#Playback_status).
 enum PlayerState {
+  /// Denotes State when player is not loaded with video.
   unknown,
+
+  /// Denotes state when player loads first video.
   unStarted,
+
+  /// Denotes state when player ends playing a video.
+  ///
+  /// Fires only once after video ends and switches to [PlayerState.stopped].
   ended,
+
+  /// Denotes state when player is playing video.
   playing,
+
+  /// Denotes state when player is paused.
   paused,
+
+  /// Denotes state when player is buffering bytes from the internet.
   buffering,
+
+  /// Denotes state when player loads video and is ready to be played.
   cued,
+
+  /// Denotes state when player has stopped playing videos.
   stopped,
 }