+++ lib/src/enums/thumbnail_quality.dart
@@ -4,9 +4,18 @@
 
 /// Quality of YouTube video thumbnail.
 class ThumbnailQuality {
-  static const String defaultQuality = 'default.jpg';
-  static const String high = 'hqdefault.jpg';
-  static const String medium = 'mqdefault.jpg';
-  static const String standard = 'sddefault.jpg';
-  static const String max = 'maxresdefault.jpg';
+  /// 120*90
+  static const String defaultQuality = 'default.webp';
+
+  /// 320*180
+  static const String medium = 'mqdefault.webp';
+
+  /// 480*360
+  static const String high = 'hqdefault.webp';
+
+  /// 640*480
+  static const String standard = 'sddefault.webp';
+
+  /// Unscaled thumbnail
+  static const String max = 'maxresdefault.webp';
 }