+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 6.0.1
+* **(New Feature)** Added `title` and `author` property to `YoutubePlayerController`.
+* Removed **DataConnectionChecker** dependency.
+* Removed `start` and `end` from flags as it wasn't functioning anyway.
+* **(Improvement)** Implemented effective dart lints.
+
 ## 6.0.0
 * **(Improvement)** Smooth fullscreen toggle and crash fix. [Issue #46](https://github.com/sarbagyastha/youtube_player_flutter/issues/46) & [Issue #105](https://github.com/sarbagyastha/youtube_player_flutter/issues/105)
 * **(Fixed)** [Issue #93](https://github.com/sarbagyastha/youtube_player_flutter/issues/93).